+++ PayPalNativeCheckout/src/main/java/com/braintreepayments/api/PayPalNativeCheckoutAccount.java
@@ -69,7 +69,7 @@ class PayPalNativeCheckoutAccount extends PaymentMethod {
      * Used by PayPal wrappers to construct a request to create a PayPal account.
      *
      * @param clientMetadataId Application clientMetadataId created by
-     *                         {@link PayPalDataCollector#getClientMetadataId(Context, Configuration)}.
+     *                         {@link DataCollector#getClientMetadataId(Context, Configuration)}.
      */
     void setClientMetadataId(String clientMetadataId) {
         this.clientMetadataId = clientMetadataId;