+++ PayPalNativeCheckout/src/main/java/com/braintreepayments/api/PayPalNativeCheckoutInternalClient.java
@@ -15,23 +15,20 @@ class PayPalNativeCheckoutInternalClient {
             "paypal_hermes/setup_billing_agreement";
 
     private final BraintreeClient braintreeClient;
-    private final PayPalDataCollector payPalDataCollector;
+    private final DataCollector dataCollector;
     private final ApiClient apiClient;
 
     private final String cancelUrl;
     private final String successUrl;
 
     PayPalNativeCheckoutInternalClient(BraintreeClient braintreeClient) {
-        this(braintreeClient, new PayPalDataCollector(braintreeClient),
-                new ApiClient(braintreeClient));
+        this(braintreeClient, new DataCollector(braintreeClient), new ApiClient(braintreeClient));
     }
 
     @VisibleForTesting
-    PayPalNativeCheckoutInternalClient(BraintreeClient braintreeClient,
-                                       PayPalDataCollector payPalDataCollector,
-                                       ApiClient apiClient) {
+    PayPalNativeCheckoutInternalClient(BraintreeClient braintreeClient, DataCollector dataCollector, ApiClient apiClient) {
         this.braintreeClient = braintreeClient;
-        this.payPalDataCollector = payPalDataCollector;
+        this.dataCollector = dataCollector;
         this.apiClient = apiClient;
         this.cancelUrl =
                 String.format("%s://onetouch/v1/cancel", braintreeClient.getReturnUrlScheme());
@@ -71,15 +68,10 @@ class PayPalNativeCheckoutInternalClient {
                                     if (redirectUrl != null) {
                                         Uri parsedRedirectUri = Uri.parse(redirectUrl);
 
-                                        String pairingIdKey =
-                                                isBillingAgreement ? "ba_token" : "token";
-                                        String pairingId =
-                                                parsedRedirectUri.getQueryParameter(pairingIdKey);
-                                        String clientMetadataId =
-                                                payPalRequest.getRiskCorrelationId() != null
-                                                        ? payPalRequest.getRiskCorrelationId() :
-                                                        payPalDataCollector.getClientMetadataId(
-                                                                context, configuration);
+                                        String pairingIdKey = isBillingAgreement ? "ba_token" : "token";
+                                        String pairingId = parsedRedirectUri.getQueryParameter(pairingIdKey);
+                                        String clientMetadataId = payPalRequest.getRiskCorrelationId() != null
+                                                ? payPalRequest.getRiskCorrelationId() : dataCollector.getClientMetadataId(context, configuration);
 
                                         if (pairingId != null) {
                                             payPalResponse