+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Braintree Android SDK Release Notes
 
-## 4.39.0 (2023-10-16)
+## unreleased
 
 * Breaking Changes
   * All Modules
@@ -9,6 +9,11 @@
   * UnionPay
     * Remove `union-pay` module
       * UnionPay cards can now be processed as regular cards (through the `card` module) due to their partnership with Discover
+  * BraintreeDataCollector
+    * Replace `DataCollector#collectDeviceData(context, merchantId, callback)` with 
+      `DataCollector#collectDeviceData(context, riskCorrelationId, callback)`
+  * PayPalDataCollector
+    * Remove `paypal-data-collector` module (use `data-collector`)
   * Venmo
     * Remove `VenmoListener`, `VenmoTokenizeAccountCallback`
     * Add `VenmoLauncher`, `VenmoAuthChallenge`, `VenmoAuthChallengeCallback`, 
@@ -33,7 +38,7 @@
       `PayPalClient`
     * Change `PayPalClient#tokenizePayPalAccount` parameters
 
-## unreleased
+## 4.39.0 (2023-10-16)
 
 * BraintreeCore
   * Remove beta features `PreferredPaymentMethodsClient`, `PreferredPaymentMethodsResult`, and `PreferredPaymentmethodsCallback`